Enhances BenefitPay integration and fixes deep link handling
Improves BenefitPay integration by allowing the inspection of the WebView for debugging purposes on iOS 16.4 and later.
It also fixes the deep link handling to properly open the BenefitPay app when a custom scheme is detected.
Additionally, configures the WebView for mobile content mode and modifies auto dismiss in example app.